+++ internal/bundler/bundler.go
@@ -436,7 +436,7 @@ func parseFile(args parseArgs) {
 				}
 
 				// Run the resolver and log an error if the path couldn't be resolved
-				resolveResult, didLogError, notes := runOnResolvePlugins(
+				resolveResult, didLogError, debug := runOnResolvePlugins(
 					args.options.Plugins,
 					args.res,
 					args.log,
@@ -484,7 +484,7 @@ func parseFile(args parseArgs) {
 							hint = fmt.Sprintf(" (the plugin %q didn't set a resolve directory)", pluginName)
 						}
 						args.log.AddRangeErrorWithNotes(&source, record.Range,
-							fmt.Sprintf("Could not resolve %q%s", record.Path.Text, hint), notes)
+							fmt.Sprintf("Could not resolve %q%s", record.Path.Text, hint), debug.Notes(&source, record.Range))
 					}
 					continue
 				}
@@ -652,7 +652,7 @@ func runOnResolvePlugins(
 	kind ast.ImportKind,
 	absResolveDir string,
 	pluginData interface{},
-) (*resolver.ResolveResult, bool, []logger.MsgData) {
+) (*resolver.ResolveResult, bool, resolver.DebugMeta) {
 	resolverArgs := config.OnResolveArgs{
 		Path:       path,
 		ResolveDir: absResolveDir,
@@ -681,7 +681,7 @@ func runOnResolvePlugins(
 
 			// Stop now if there was an error
 			if didLogError {
-				return nil, true, nil
+				return nil, true, resolver.DebugMeta{}
 			}
 
 			// The "file" namespace is the default for non-external paths, but not
@@ -710,21 +710,21 @@ func runOnResolvePlugins(
 					log.AddRangeError(importSource, importPathRange,
 						fmt.Sprintf("Plugin %q returned a non-absolute path: %s (set a namespace if this is not a file path)", pluginName, result.Path.Text))
 				}
-				return nil, true, nil
+				return nil, true, resolver.DebugMeta{}
 			}
 
 			return &resolver.ResolveResult{
 				PathPair:   resolver.PathPair{Primary: result.Path},
 				IsExternal: result.External,
 				PluginData: result.PluginData,
-			}, false, nil
+			}, false, resolver.DebugMeta{}
 		}
 	}
 
 	// Resolve relative to the resolve directory by default. All paths in the
 	// "file" namespace automatically have a resolve directory. Loader plugins
 	// can also configure a custom resolve directory for files in other namespaces.
-	result, notes := res.Resolve(absResolveDir, path, kind)
+	result, debug := res.Resolve(absResolveDir, path, kind)
 
 	// Warn when the case used for importing differs from the actual file name
 	if result != nil && result.DifferentCase != nil && !resolver.IsInsideNodeModules(absResolveDir) {
@@ -736,7 +736,7 @@ func runOnResolvePlugins(
 		))
 	}
 
-	return result, false, notes
+	return result, false, debug
 }
 
 type loaderPluginResult struct {
@@ -1202,7 +1202,7 @@ func (s *scanner) addEntryPoints(entryPoints []string) []uint32 {
 	for i, path := range entryPoints {
 		go func(i int, path string) {
 			// Run the resolver and log an error if the path couldn't be resolved
-			resolveResult, didLogError, notes := runOnResolvePlugins(
+			resolveResult, didLogError, debug := runOnResolvePlugins(
 				s.options.Plugins,
 				s.res,
 				s.log,
@@ -1227,7 +1227,7 @@ func (s *scanner) addEntryPoints(entryPoints []string) []uint32 {
 						hint = fmt.Sprintf(" (use %q to reference the file %q)", "./"+path, s.res.PrettyPath(query.PathPair.Primary))
 					}
 				}
-				s.log.AddErrorWithNotes(nil, logger.Loc{}, fmt.Sprintf("Could not resolve %q%s", path, hint), notes)
+				s.log.AddErrorWithNotes(nil, logger.Loc{}, fmt.Sprintf("Could not resolve %q%s", path, hint), debug.Notes(nil, logger.Range{}))
 			}
 			entryPointWaitGroup.Done()
 		}(i, path)